[BC] Break down the plib_git_dirty function
The 'plib_git_dirty' function performs multiple tasks at once and this
hurts re-usability and customization possibilities.

In short, it breaks the single responsibility principle.

We break it down into:
- plib_git_status: it returns the git status --porcelain return value.
- plib_git_staged_mod: returns the number of staged file modifications.
- plib_git_unstaged_mod: returns the number of unstaged file
  modifications.
- plib_git_staged_del: returns the number of staged file deletions.
- plib_git_unstaged_del: returns the number of unstaged file deletions.
- plib_git_staged_add: returns the number of staged file additions.
- plib_git_unstaged_add: returns the number of unstaged file additions.
  (I don't know how to produce this scenario tbh).
- plib_git_status_new: returns the number of untracked files.

All those functionalities are taken from the original 'plib_git_dirty'.
There are no new functionality.

Everything related to formating has been removed. It should be the
responsibility of the calling script to format data as it pleases.

The objective is to make this library purely data focused. It should not
handle any kind of formating.

"Tests" have been updated to match the new implementation.